Initial commit
 Changes to be committed:
	new file:   GAAmainmapper.js
	new file:   index.html
	new file:   init script modify for system
	new file:   old index.html
	new file:   pageactions.js
	new file:   print.css
	new file:   startscript

This program is pretty simple. It just takes a google mapsengine link and
makes it displayable within an iframe that is generated for the user. This
was done because the client was having issues printing the maps from
google at the zoom level they needed in order to route drivers. Putting
the map into an iframe solved this issue.